\name{example}

\alias{example}

\alias{d}

\title{Example network}

\usage{

data(example)

}

\description{

Example data set: synthetic, undirected network with 15 nodes.

}

\value{Undirected network.}

\seealso{network, hergm, ergm.terms, hergm.terms}

\references{

Schweinberger, M. and P. Luna (2018). HERGM: Hierarchical exponential-family random graph models. Journal of Statistical Software, 85, 1--39.

}
