\title{Internal hergm function}

\name{hergm-internal}

\alias{hergm.internal}

\alias{compute_initial_parameters}

\alias{compute_initial_parameters_curved}

\alias{check_extensions} 

\alias{length_mcmc}

\alias{number_clusters}

\alias{bernoulli_map_mean_to_natural}

\alias{hergm.initialize}

\alias{hergm.loss}

\alias{hergm.mcmc}

\alias{hergm.min_loss_1} 

\alias{hergm.min_loss_2} 

\alias{hergm.p}

\alias{hergm.relabel_1}

\alias{hergm.relabel_2}

\alias{hergm.permute} 

\alias{hergm.permute_indicator}

\alias{hergm.permute_mcmc}

\alias{hergm.pie}

\alias{hergm.preprocess} 

\alias{hergm.sa} 

\alias{hergm.set.mcmc} 

\alias{hergm.step_1} 

\alias{hergm.step_2}

\alias{mcmc.diagnostics.hergm}

\alias{summary_sample_network}

\alias{hergm.wrapper} 

\alias{hergm.permutation.wrapper} 

\alias{InitErgmTerm.ctriple_ijk} 

\alias{InitErgmTerm.edges_i} 

\alias{InitErgmTerm.arcs_i} 

\alias{InitErgmTerm.arcs_j} 

\alias{InitErgmTerm.edges_ij}

\alias{InitErgmTerm.mutual_i} 

\alias{InitErgmTerm.mutual_ij} 

\alias{InitErgmTerm.transitiveties_ijk} 

\alias{InitErgmTerm.triangle_ijk} 

\alias{InitErgmTerm.ttriple_ijk}

\alias{InitErgmTerm.twostar_ijk}

\alias{hergm.getnetwork}

\alias{plot_weighted_graph}

\alias{silent}

\alias{step_to_chull}

\alias{opt_fisher}

\alias{comp_weights}

\alias{comp_info_mat}

\alias{comp_info_mat_par}

\alias{comp_step}

\alias{check_convergence}

\alias{MCMLE}

\alias{MCMC_sample}

\alias{MCMC_sample_par}

\alias{est_ergm}

\alias{est_control}

\alias{initialize_object}

\alias{permute_tau}

\alias{logit}

\alias{inv.logit}

\alias{spec_clust}

\alias{check_clusters}

\alias{simulate_hergm}

\alias{init_estimate_params}

\alias{estimate_params}

\alias{estimate_bw_param}

\alias{estimate_density}

\alias{EM_wrapper_fast}

\alias{hergm.large}

\alias{assign_labels}

\alias{calculateStats}

\alias{dim_fun}

\alias{easy_E_step}

\alias{easy_M_Step}

\alias{exp_fun}

\alias{isTauSignificantlyChanged}

\alias{log_fun}

\alias{make_eta_fun}

\alias{make_eta_grad}

\alias{make_return_obj}

\alias{onLoad}

\alias{outer_fun}

\alias{par_sim_fun}

\alias{rep_row}

\alias{runFixedPointEstimationEStep}

\alias{runFixedPointEstimationEStepMM}

\alias{runModelEstimationMStep}

\alias{seq_spec}

\alias{sort_neighborhoods}

\alias{split_neighborhoods}

\alias{same_clusters}

\alias{plot_neighborhoods}

\alias{find_phi_coefficient}

\alias{par_get_subgraph}

\alias{norm1}

\alias{norm2}

\alias{get_coef_names} 

\alias{nvl_mimic}

\alias{ergm.Cprepare}

\description{

Internal hergm function.

}

\details{

Internal function which is called by user-level function.

}

\seealso{User-level functions: hergm, hergm.terms, hergm.gof, hergm.summary}

\keyword{internal}
