
\name{hergm.plot}

\alias{hergm.plot}

\alias{plot.hergm}

\alias{plot}

\title{Plot network and neighborhoods}

\description{

The function \code{hergm.plot} plots the network and neighborhoods by exploiting posterior classification probabilities based on a MCMC sample generated by function \code{hergm}.

}

\usage{

hergm.plot(sample = NULL,
           \dots)

}

\arguments{

\item{sample}{MCMC sample generated by function \code{hergm}.}

\item{\dots}{additional arguments, to be passed to lower-level functions in the future.}

}

\value{

Plot of network and neighborhoods.

}

\seealso{network, hergm}

\examples{
\dontrun{
data(example)

sample <- hergm(d ~ edges_ij + triangle_ijk)

hergm.plot(sample)
}
}

