\name{summary}

\alias{summary}

\alias{summary.hergm}

\alias{hergm.summary}

\title{Summary}

\description{

The function \code{summary.hergm} produces summaries of the MCMC sample.

}

\usage{

\method{summary}{hergm}(sample = NULL,
        \dots)

}

\arguments{

\item{sample}{MCMC sample generated by function \code{hergm}.}

\item{\dots}{additional arguments, to be passed to lower-level functions in the future.}

}

\value{

Summary function.

}

\seealso{network, hergm, hergm.plot}

\examples{
\dontrun{
data(example)

sample <- hergm(d ~ edges_ij + triangle_ijk)

summary.hergm(sample)
}
}

