\name{kapferer}

\alias{kapferer}

\title{Kapferer collaboration network}

\usage{

data(kapferer)

}

\description{

The network corresponds to collaborations between 39 workers in a tailor shop in Africa:
an undirected edge between workers \code{i} and \code{j} indicates that the workers collaborated.
The network is taken from Kapferer (1972).

}

\value{Undirected network.}

\seealso{network, hergm, ergm.terms, hergm.terms}

\references{

Kapferer, B. (1972). Strategy and Transaction in an African Factory. Manchester University
Press, Manchester, U.K.

Schweinberger, M. and P. Luna (2018). HERGM: Hierarchical exponential-family random graph models. Journal of Statistical Software, 85, 1--39.

}
