\title{Internal hergm function}

\name{hergm-internal}

\alias{hergm.internal}

\alias{length_mcmc}

\alias{number_clusters}

\alias{bernoulli_map_mean_to_natural}

\alias{hergm.initialize}

\alias{hergm.loss}

\alias{hergm.mcmc}

\alias{hergm.min_loss_1} 

\alias{hergm.min_loss_2} 

\alias{hergm.p}

\alias{hergm.relabel_1}

\alias{hergm.relabel_2}

\alias{hergm.permute} 

\alias{hergm.permute_indicator}

\alias{hergm.permute_mcmc}

\alias{hergm.pie}

\alias{hergm.preprocess} 

\alias{hergm.sa} 

\alias{hergm.set.mcmc} 

\alias{hergm.step_1} 

\alias{hergm.step_2}

\alias{summary_sample_network}

\alias{hergm.wrapper} 

\alias{hergm.permutation.wrapper} 

\alias{InitErgmTerm.ctriple_ijk} 

\alias{InitErgmTerm.edges_i} 

\alias{InitErgmTerm.arcs_i} 

\alias{InitErgmTerm.arcs_j} 

\alias{InitErgmTerm.edges_ij}

\alias{InitErgmTerm.mutual_i} 

\alias{InitErgmTerm.mutual_ij} 

\alias{InitErgmTerm.triangle_ijk} 

\alias{InitErgmTerm.ttriple_ijk}

\alias{InitErgmTerm.twostar_ijk}

\alias{hergm.getnetwork}

\description{

Internal hergm function.

}

\details{

Internal function which is called by user-level function.

}

\seealso{User-level functions: hergm, hergm.terms, hergm.gof, hergm.summary}

\keyword{internal}

