
\name{hergm-internal}

\alias{hergm.internal}

\title{Internal hergm function}

\alias{length_mcmc}

\alias{number_clusters}

\alias{bernoulli_map_mean_to_natural}

\alias{hergm.loss}

\alias{hergm.mcmc}

\alias{hergm.min_loss} 

\alias{hergm.p}

\alias{hergm.permute} 

\alias{hergm.permute_indicator}

\alias{hergm.permute_mcmc}

\alias{hergm.pie}

\alias{hergm.preprocess} 

\alias{hergm.sa} 

\alias{hergm.set.mcmc} 

\alias{hergm.step_1} 

\alias{hergm.step_2}

\alias{hergm.wrapper} 

\alias{InitErgm.ctriple_ijk} 

\alias{InitErgm.edges_i} 

\alias{InitErgm.arcs_i} 

\alias{InitErgm.arcs_j} 

\alias{InitErgm.edges_ij}

\alias{InitErgm.mutual_i} 

\alias{InitErgm.mutual_ij} 

\alias{InitErgm.triangle_ijk} 

\alias{InitErgm.ttriple_ijk}

\alias{InitErgm.twostar_i} 

\alias{InitErgm.twostar_ijk}

\description{

  Internal hergm function.

}

\details{

Internal function which is called by user-level function.

}

\seealso{User-level functions: hergm, hergm.terms, hergm.postprocess}

\keyword{internal}

