\name{summary.hergm}

\alias{summary.hergm}

\alias{hergm.summary}

\title{Summary of object of class \code{hergm}}

\description{

The function \code{summary.hergm} generates a summary of an object of class \code{hergm} by using the functions \code{print.hergm} and \code{plot.hergm}.
The function \code{print.hergm} prints a summary of a sample of parameters from the posterior,
whereas the function \code{plot.hergm} plots a summary of a sample of block memberships of nodes from the posterior.

}

\usage{

\method{summary}{hergm}(object, \dots)

}

\arguments{

\item{object}{object of class \code{hergm}; objects of class \code{hergm} can be generated by function \code{hergm}.}

\item{\dots}{additional arguments, to be passed to lower-level functions in the future.}

}

\seealso{hergm, hergm.postprocess, print.hergm, plot.hergm}

\references{

Schweinberger, M. and P. Luna (2018). HERGM: Hierarchical exponential-family random graph models. Journal of Statistical Software, 85, 1--39.

}

\examples{
\dontrun{data(example)
object <- hergm(d ~ edges_ij + triangle_ijk)
summary(object)
}
}

