\name{summary}

\alias{summary}

\alias{summary.hergm}

\alias{hergm.summary}

\title{Summary of object of class \code{hergm}}

\description{

The function \code{summary} generates a summary of an object of class \code{hergm} by using the functions \code{print} and \code{plot}.
The function \code{print} prints a summary of a sample of parameters from the posterior,
whereas the function \code{plot} plots a summary of a sample of block memberships of nodes from the posterior.

}

\usage{

\method{summary}{hergm}(object, \dots)

}

\arguments{

\item{object}{object of class \code{hergm}; objects of class \code{hergm} can be generated by function \code{hergm}.}

\item{\dots}{additional arguments, to be passed to lower-level functions in the future.}

}

\seealso{hergm, hergm.postprocess, print, plot}

\examples{
\dontrun{data(example)
object <- hergm(d ~ edges_ij + triangle_ijk)
summary(object)
}
}

